requirements.txt
requests>=2.32
pandas>=2.2
python-docx>=1.1
pdfminer.six>=20231228
openpyxl>=3.1
PyYAML>=6.0
python-dotenv>=1.0
tenacity>=8.4
tqdm>=4.66
typer>=0.12
rich>=13.9
streamlit>=1.38
pydantic>=2.8

# Optional OCR/conversion extras
# pytesseract>=0.3
# pillow>=10
